refactor: move some functions (for work with local storage, for save the utility's result to file) to new modules
