+++ rss_reader/__init__.py
@@ -1,2 +1,1 @@
 import rss_reader.rss_reader
-import rss_reader.tests
