+++ rss_reader/rss_reader/__init__.py
@@ -1,6 +1,1 @@
-from .rss_reader import main, RSSReader
-
-__all__ = [
-    'main',
-    'RSSReader',
-]
+from .rss_reader import main
