--- rss_reader/tests/__init__.py
@@ -1,3 +1,0 @@
-from .test_rss_reader import RSSReaderTests
-
-__all__ = ['RSSReaderTests']
