+++ rss_reader/tests/test_storage.json
@@ -1,1 +1,26 @@
-[{"channel_id": "https://news.ru/", "channel_title": "News channel", "news": [{"number": 1, "title": "Last news", "link": "https://news.ru/last_news.html", "author": "Alexey Morozov", "date": "Mon, 24 May, 2021 10:14 AM", "image": "http://test.com/image.jpg", "description": "Last news"}, {"number": 2, "title": "Next news", "link": "https://super_news.ru/next_news.html", "author": "", "date": "", "image": "", "description": ""}]}]
+[
+    {
+        "channel_id": "https://news.ru/",
+        "channel_title": "News channel",
+        "news": [
+            {
+                "number": 1,
+                "title": "Last news",
+                "link": "https://news.ru/last_news.html",
+                "author": "Alexey Morozov",
+                "date": "Mon, 24 May, 2021 10:14 AM",
+                "image": "http://test.com/image.jpg",
+                "description": "Last news"
+            },
+            {
+                "number": 2,
+                "title": "Next news",
+                "link": "https://super_news.ru/next_news.html",
+                "author": "",
+                "date": "",
+                "image": "",
+                "description": ""
+            }
+        ]
+    }
+]
